# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "metawall API",
    "description": "metawall API文件"
  },
  "host": "https://secure-plains-31314.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "authorization",
      "description": "請加上 API Token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/posts": {
      "get": {
        "tags": [
          "Posts - 貼文"
        ],
        "description": "取得全部貼文 API",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "Bearer token"
          },
          {
            "name": "timeSort",
            "in": "query",
            "type": "string",
            "required": false,
            "description": "排序：desc/asc"
          },
          {
            "name": "q",
            "in": "query",
            "type": "string",
            "required": false,
            "description": "關鍵字搜尋"
          }
        ],
        "responses": {
          "200": {
            "description": "貼文資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "6278da02631ef7910e7adc37"
                      },
                      "content": {
                        "type": "string",
                        "example": "今天天氣真好"
                      },
                      "image": {
                        "type": "string",
                        "example": ""
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "6277b20ad980d4df45db6447"
                          },
                          "name": {
                            "type": "string",
                            "example": "Johnny"
                          },
                          "photo": {
                            "type": "string",
                            "example": ""
                          }
                        }
                      },
                      "likes": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Posts - 貼文"
        ],
        "description": "刪除全部貼文 API",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "Bearer token"
          }
        ],
        "responses": {
          "200": {
            "description": "貼文資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "object",
                  "properties": {
                    "acknowledged": {
                      "type": "boolean",
                      "example": true
                    },
                    "deletedCount": {
                      "type": "number",
                      "example": 10
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}